+++ stdlib/area/position.lua
@@ -5,6 +5,26 @@ Position = {}
 
 require 'stdlib/core'
 
+
+--- Creates a table representing the position from x and y
+-- @param x x-position
+-- @param y y-position
+-- @return Position
+function Position.construct(x, y)
+    fail_if_missing(x, "missing x position argument")
+    fail_if_missing(y, "missing y position argument")
+    return { x = x, y = y }
+end
+
+--- Creates a position that is a copy of the given position
+-- @param pos the position to copy
+-- @return Position
+function Position.copy(pos)
+    fail_if_missing(pos, "missing position argument")
+    pos = Position.to_table(pos)
+    return { x = pos.x, y = pos.y }
+end
+
 --- Creates a position that is offset by x,y coordinate pair
 -- @param pos the position to offset
 -- @param x the amount to offset the position in the x direction