+++ stdlib/game.lua
@@ -9,14 +9,16 @@ end
 --- Messages all players currently connected to the game
 -- @param msg message to send to players
 -- @param condition (optional) optional condition to be true for the player to be messaged
--- @return the number of players who received the message
+-- @return the number of players who received the message. Offline players are not counted as having received the message.
 function Game.print_all(msg, condition)
     local num = 0
     for _, player in pairs(game.players) do
-        if player.valid and player.connected then
+        if player.valid then
             if condition == nil or select(2, pcall(condition, player)) then
                 player.print(msg)
-                num = num + 1
+                if player.connected then
+                    num = num + 1
+                end
             end
         end
     end
@@ -24,6 +26,7 @@ function Game.print_all(msg, condition)
 end
 
 --- Messages all players with the given force connected to the game
+-- <b>Deprecated for Factorio 0.14+</b>, see force.print(msg) instead.
 -- @param force (may be force name string, or force object) the players with the given force to message
 -- @param msg message to send to players
 -- @return the number of players who received the message
@@ -40,6 +43,7 @@ function Game.print_force(force, msg)
 end
 
 --- Messages all players with the given surface connected to the game
+-- <b>Deprecated for Factorio 0.14+</b>, see surface.print(msg) instead.
 -- @param surface the players with the given surface to message
 -- @param msg message to send to players
 -- @return the number of players who received the message